requirements.txt
fastapi
httpx
requests
uvicorn
python-dotenv
redis
pydantic-settings
aiogram
tenacity